\`Refactor\`: Resolve runtime issues from new document model migration
Fixes the test suite to fully pass on the new model. Five issues
addressed:

1. Bean name conflicts: Spring auto-derived "documentService",
   "documentResource", and "documentRepository" from class simple names,
   but legacy and new versions of each shared the same simple name and
   collided. Added explicit bean names "legacyDocumentService",
   "legacyDocumentResource", and "legacyDocumentRepository" on the
   legacy classes (they get deleted in session 3).

2. ArchUnit data-export rule: ApplicantDocument and ApplicationDocument
   now carry @NoUserDataExportRequired, matching the base.

3. Liquibase index too long: shrunk documents_v2.path from VARCHAR(1024)
   to VARCHAR(512) so the path index fits MySQL's 3072-byte key limit
   under utf8mb4.

4. Orphan-file cleanup leniency: DocumentService.removeFileIfOrphan now
   logs and skips when a stored path lies outside the storage root
   (e.g., test fixtures using mock paths) instead of failing the
   transaction.

5. Trusted internal download path: added
   DocumentService.loadResourceForExport(Document) for the user data
   export ZIP writer, which runs as a scheduled job without a security
   context and therefore can't go through the access-checking
   downloadDocument(UUID). UserExportZipWriter now uses the new
   DocumentRepository to fetch uploaded documents, mirroring the legacy
   path's behavior. EmailServiceTest mocks updated to the new
   DocumentService API surface (downloadDocument(UUID) instead of the
   legacy two-step lookup).

Result: 747 tests, 0 failures.
