+++ src/main/java/de/tum/cit/aet/core/documents/domain/ApplicationDocument.java
@@ -1,6 +1,7 @@
 package de.tum.cit.aet.core.documents.domain;
 
 import de.tum.cit.aet.application.domain.Application;
+import de.tum.cit.aet.core.domain.export.NoUserDataExportRequired;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
@@ -14,6 +15,7 @@ import lombok.Setter;
  */
 @Entity
 @DiscriminatorValue("APPLICATION")
+@NoUserDataExportRequired(reason = "Documents are exported as binary files by UserExportZipWriter")
 @Getter
 @Setter
 public class ApplicationDocument extends Document {
