+++ src/main/java/de/tum/cit/aet/core/repository/DocumentRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-@Repository
+@Repository("legacyDocumentRepository")
 public interface DocumentRepository extends TumApplyJpaRepository<Document, UUID> {
     Optional<Document> findBySha256Id(String sha256Id);
