+++ src/main/java/de/tum/cit/aet/core/service/DocumentService.java
@@ -27,7 +27,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 
-@Service
+@Service("legacyDocumentService")
 public class DocumentService {
 
     private final DocumentRepository documentRepository;
