+++ src/main/java/de/tum/cit/aet/core/service/export/UserExportZipWriter.java
@@ -1,6 +1,7 @@
 package de.tum.cit.aet.core.service.export;
 
-import de.tum.cit.aet.core.domain.Document;
+import de.tum.cit.aet.core.documents.domain.Document;
+import de.tum.cit.aet.core.documents.repository.DocumentRepository;
 import de.tum.cit.aet.core.domain.Image;
 import de.tum.cit.aet.core.dto.exportdata.ApplicantDataExportDTO;
 import de.tum.cit.aet.core.dto.exportdata.ApplicantInternalCommentExportDTO;
@@ -8,7 +9,6 @@ import de.tum.cit.aet.core.dto.exportdata.ApplicantRatingExportDTO;
 import de.tum.cit.aet.core.dto.exportdata.StaffDataDTO;
 import de.tum.cit.aet.core.dto.exportdata.UserDataExportDTO;
 import de.tum.cit.aet.core.exception.UserDataExportException;
-import de.tum.cit.aet.core.repository.DocumentRepository;
 import de.tum.cit.aet.core.repository.ImageRepository;
 import de.tum.cit.aet.core.service.ZipExportService;
 import de.tum.cit.aet.core.util.FileUtil;
