+++ src/main/resources/config/liquibase/changelog/00000000000035_create_documents_v2_table.xml
@@ -26,7 +26,7 @@
       <column name="name" type="VARCHAR(512)">
         <constraints nullable="false"/>
       </column>
-      <column name="path" type="VARCHAR(1024)">
+      <column name="path" type="VARCHAR(512)">
         <constraints nullable="false"/>
       </column>
       <column name="mime_type" type="VARCHAR(128)">
